Split up visitor and scope filtering logic.
