+++ src/workers/parser/getScopes/index.js
@@ -4,11 +4,10 @@
 
 // @flow
 
-import { traverseAst } from "../utils/ast";
 import {
-  createParseJSScopeVisitor,
-  findScopes,
-  type SourceScope
+  parseSourceScopes,
+  type SourceScope,
+  type ParsedScope
 } from "./visitor";
 
 import type { Location } from "../../../types";
@@ -19,9 +18,7 @@ export default function getScopes(location: Location): SourceScope[] {
   const { sourceId } = location;
   let parsedScopes = parsedScopesCache.get(sourceId);
   if (!parsedScopes) {
-    const visitor = createParseJSScopeVisitor(sourceId);
-    traverseAst(sourceId, visitor.traverseVisitor);
-    parsedScopes = visitor.toParsedScopes();
+    parsedScopes = parseSourceScopes(sourceId);
     parsedScopesCache.set(sourceId, parsedScopes);
   }
   return findScopes(parsedScopes, location);
@@ -29,4 +26,46 @@ export default function getScopes(location: Location): SourceScope[] {
 
 export function clearScopes() {
   parsedScopesCache = new Map();
+}
+
+/**
+ * Searches all scopes and their bindings at the specific location.
+ */
+function findScopes(scopes: ParsedScope[], location: Location): SourceScope[] {
+  // Find inner most in the tree structure.
+  let searchInScopes: ?(ParsedScope[]) = scopes;
+  const found = [];
+  while (searchInScopes) {
+    const foundOne = searchInScopes.some(s => {
+      if (
+        compareLocations(s.start, location) <= 0 &&
+        compareLocations(location, s.end) < 0
+      ) {
+        // Found the next scope, trying to search recusevly in its children.
+        found.unshift(s);
+        searchInScopes = s.children;
+        return true;
+      }
+      return false;
+    });
+    if (!foundOne) {
+      break;
+    }
+  }
+  return found.map(i => {
+    return {
+      type: i.type,
+      displayName: i.displayName,
+      start: i.start,
+      end: i.end,
+      bindings: i.bindings
+    };
+  });
+}
+
+function compareLocations(a: Location, b: Location): number {
+  // According to type of Location.column can be undefined, if will not be the
+  // case here, ignoring flow error.
+  // $FlowIgnore
+  return a.line == b.line ? a.column - b.column : a.line - b.line;
 }
